refactor: vajra.cpp into focused units
- Included runtime_logging.cpp in the vajra_server_test executable.
- This addition enhances the logging capabilities of the server, allowing for better debugging and monitoring during tests.

closes: #387
